+++ src/main/java/com/kamijoucen/ruler/logic/function/math/SqrtFunction.java
@@ -9,6 +9,8 @@ import com.kamijoucen.ruler.domain.value.IntegerValue;
 import com.kamijoucen.ruler.domain.value.ValueType;
 import com.kamijoucen.ruler.logic.function.RulerFunction;
 
+import java.math.BigDecimal;
+
 public class SqrtFunction implements RulerFunction {
 
     @Override
@@ -24,12 +26,12 @@ public class SqrtFunction implements RulerFunction {
         BaseValue value = (BaseValue) param[0];
         double d;
         if (value.getType() == ValueType.INTEGER) {
-            d = ((IntegerValue) value).getValue();
+            d = ((IntegerValue) value).getValue().doubleValue();
         } else if (value.getType() == ValueType.DOUBLE) {
-            d = ((DoubleValue) value).getValue();
+            d = ((DoubleValue) value).getValue().doubleValue();
         } else {
             throw new RulerRuntimeException("sqrt expects a number");
         }
-        return new DoubleValue(Math.sqrt(d));
+        return new DoubleValue(BigDecimal.valueOf(Math.sqrt(d)));
     }
 }
