+++ src/test/java/com/kamijoucen/ruler/test/ScopeTest.java
@@ -7,21 +7,23 @@ import com.kamijoucen.ruler.domain.exception.RulerRuntimeException;
 import org.junit.Assert;
 import org.junit.Test;
 
+import java.math.BigInteger;
+
 public class ScopeTest {
 
     @Test
     public void findLocalTest() {
         Scope scope = new Scope("test", false, null, null);
-        scope.putLocal("a", new IntegerValue(10));
-        Assert.assertEquals(10, ((IntegerValue) scope.find("a")).getValue());
+        scope.putLocal("a", new IntegerValue(BigInteger.valueOf(10)));
+        Assert.assertEquals(BigInteger.valueOf(10), ((IntegerValue) scope.find("a")).getValue());
     }
 
     @Test
     public void findParentTest() {
         Scope parent = new Scope("parent", false, null, null);
-        parent.putLocal("a", new IntegerValue(20));
+        parent.putLocal("a", new IntegerValue(BigInteger.valueOf(20)));
         Scope child = new Scope("child", false, parent, null);
-        Assert.assertEquals(20, ((IntegerValue) child.find("a")).getValue());
+        Assert.assertEquals(BigInteger.valueOf(20), ((IntegerValue) child.find("a")).getValue());
     }
 
     @Test
@@ -33,54 +35,54 @@ public class ScopeTest {
     @Test
     public void shadowingTest() {
         Scope parent = new Scope("parent", false, null, null);
-        parent.putLocal("a", new IntegerValue(1));
+        parent.putLocal("a", new IntegerValue(BigInteger.valueOf(1)));
         Scope child = new Scope("child", false, parent, null);
-        child.putLocal("a", new IntegerValue(2));
-        Assert.assertEquals(2, ((IntegerValue) child.find("a")).getValue());
-        Assert.assertEquals(1, ((IntegerValue) parent.find("a")).getValue());
+        child.putLocal("a", new IntegerValue(BigInteger.valueOf(2)));
+        Assert.assertEquals(BigInteger.valueOf(2), ((IntegerValue) child.find("a")).getValue());
+        Assert.assertEquals(BigInteger.valueOf(1), ((IntegerValue) parent.find("a")).getValue());
     }
 
     @Test
     public void updateLocalTest() {
         Scope scope = new Scope("test", false, null, null);
-        scope.putLocal("a", new IntegerValue(1));
-        scope.update("a", new IntegerValue(99));
-        Assert.assertEquals(99, ((IntegerValue) scope.find("a")).getValue());
+        scope.putLocal("a", new IntegerValue(BigInteger.valueOf(1)));
+        scope.update("a", new IntegerValue(BigInteger.valueOf(99)));
+        Assert.assertEquals(BigInteger.valueOf(99), ((IntegerValue) scope.find("a")).getValue());
     }
 
     @Test
     public void updateParentTest() {
         Scope parent = new Scope("parent", false, null, null);
-        parent.putLocal("a", new IntegerValue(1));
+        parent.putLocal("a", new IntegerValue(BigInteger.valueOf(1)));
         Scope child = new Scope("child", false, parent, null);
-        child.update("a", new IntegerValue(99));
-        Assert.assertEquals(99, ((IntegerValue) parent.find("a")).getValue());
+        child.update("a", new IntegerValue(BigInteger.valueOf(99)));
+        Assert.assertEquals(BigInteger.valueOf(99), ((IntegerValue) parent.find("a")).getValue());
     }
 
     @Test(expected = RulerRuntimeException.class)
     public void updateMissingTest() {
         Scope scope = new Scope("test", false, null, null);
-        scope.update("missing", new IntegerValue(1));
+        scope.update("missing", new IntegerValue(BigInteger.valueOf(1)));
     }
 
     @Test
     public void defineLocalTest() {
         Scope scope = new Scope("test", false, null, null);
-        scope.defineLocal("a", new IntegerValue(5));
-        Assert.assertEquals(5, ((IntegerValue) scope.getByLocal("a")).getValue());
+        scope.defineLocal("a", new IntegerValue(BigInteger.valueOf(5)));
+        Assert.assertEquals(BigInteger.valueOf(5), ((IntegerValue) scope.getByLocal("a")).getValue());
     }
 
     @Test(expected = RulerRuntimeException.class)
     public void defineLocalDuplicateTest() {
         Scope scope = new Scope("test", false, null, null);
-        scope.defineLocal("a", new IntegerValue(1));
-        scope.defineLocal("a", new IntegerValue(2));
+        scope.defineLocal("a", new IntegerValue(BigInteger.valueOf(1)));
+        scope.defineLocal("a", new IntegerValue(BigInteger.valueOf(2)));
     }
 
     @Test
     public void getByLocalOnlyTest() {
         Scope parent = new Scope("parent", false, null, null);
-        parent.putLocal("a", new IntegerValue(1));
+        parent.putLocal("a", new IntegerValue(BigInteger.valueOf(1)));
         Scope child = new Scope("child", false, parent, null);
         Assert.assertNull(child.getByLocal("a"));
     }
@@ -88,7 +90,7 @@ public class ScopeTest {
     @Test
     public void removeTest() {
         Scope scope = new Scope("test", false, null, null);
-        scope.putLocal("a", new IntegerValue(1));
+        scope.putLocal("a", new IntegerValue(BigInteger.valueOf(1)));
         scope.remove("a");
         Assert.assertNull(scope.find("a"));
     }
@@ -96,11 +98,11 @@ public class ScopeTest {
     @Test
     public void removeOnlyLocalTest() {
         Scope parent = new Scope("parent", false, null, null);
-        parent.putLocal("a", new IntegerValue(1));
+        parent.putLocal("a", new IntegerValue(BigInteger.valueOf(1)));
         Scope child = new Scope("child", false, parent, null);
-        child.putLocal("a", new IntegerValue(2));
+        child.putLocal("a", new IntegerValue(BigInteger.valueOf(2)));
         child.remove("a");
-        Assert.assertEquals(1, ((IntegerValue) child.find("a")).getValue());
+        Assert.assertEquals(BigInteger.valueOf(1), ((IntegerValue) child.find("a")).getValue());
     }
 
     @Test
